Add cut line behavior and missing shortcut keys.
Nano and emacs both have a cut line feature (Ctrl+K for nano) where it cuts
the current line to the clipboard or the selection. It's a mix of selecting
plus Ctrl+X, or only Ctrl+X if a selection exists already), except that you
don't have to move your cursor while selecting (it's very useful in
practice, since you only press one key instead of at least 4)

This pull request adds this feature, and a shortcut Ctrl+E to the new
document_list feature I find very useful, but inaccessible in the menu.