Jump to the matching brace using `Ctrl+J`
I'm not very proud of this code since I'm splitting the text in chars for searching and doing the search via the scripting language.
It would have been better if an `index_of` and `index_of_reverse` function existed in C for the string. Using `seek` on the string scalar segfaults and even if it doesn't, it would not work for searching backward.
Maybe it's time to add the missing functions to the scripting language ?