New action 'insert_raw_string'.
This new action and its associated keycode 'insert-raw-string' inserts
the string stored in 'mp.raw_string' at the cursor position. It's used
to bypass the actions and to insert strings directly from drivers. The
ANSI driver uses it to insert text read from stdin that is not recognized
as an ANSI sequence (to allow, among other things, pasting text when
run inside virtual terminals).